+++ FacebookAudienceNetwork/FacebookInterstitialCustomEvent.m
@@ -93,12 +93,12 @@
         MPLogAdEvent([MPLogEvent adShowAttemptForAdapter:NSStringFromClass(self.class)], self.fbPlacementId);
 
         MPLogAdEvent([MPLogEvent adWillAppearForAdapter:NSStringFromClass(self.class)], self.fbPlacementId);
-        [self.delegate fullscreenAdAdapterAdWillAppear:self];
+        [self.delegate fullscreenAdAdapterAdWillPresent:self];
 
         [self.fbInterstitialAd showAdFromRootViewController:viewController];
         
         MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], self.fbPlacementId);
-        [self.delegate fullscreenAdAdapterAdDidAppear:self];
+        [self.delegate fullscreenAdAdapterAdDidPresent:self];
         
         [self cancelExpirationTimer];
     }
@@ -192,17 +192,13 @@
 {
     MPLogAdEvent([MPLogEvent adDidDisappearForAdapter:NSStringFromClass(self.class)], self.fbPlacementId);
     [self.delegate fullscreenAdAdapterAdDidDisappear:self];
-    
-    // Signal that the fullscreen ad is closing and the state should be reset.
-    // `fullscreenAdAdapterAdDidDismiss:` was introduced in MoPub SDK 5.15.0.
-    if ([self.delegate respondsToSelector:@selector(fullscreenAdAdapterAdDidDismiss:)]) {
-        [self.delegate fullscreenAdAdapterAdDidDismiss:self];
-    }
+    [self.delegate fullscreenAdAdapterAdDidDismiss:self];
 }
 
 - (void)interstitialAdWillClose:(FBInterstitialAd *)interstitialAd
 {
     MPLogAdEvent([MPLogEvent adWillDisappearForAdapter:NSStringFromClass(self.class)], self.fbPlacementId);
+    [self.delegate fullscreenAdAdapterAdWillDismiss:self];
     [self.delegate fullscreenAdAdapterAdWillDisappear:self];
 }