+++ Pangle/PangleRewardedVideoCustomEvent.m
@@ -2,7 +2,6 @@
     #import <BUAdSDK/BUAdSDK.h>
 #if __has_include("MoPub.h")
     #import "MPLogging.h"
-    #import "MPRewardedVideoError.h"
     #import "MPReward.h"
 #endif
 #import "PangleAdapterConfiguration.h"
@@ -11,6 +10,7 @@
 @property (nonatomic, strong) BURewardedVideoAd *rewardVideoAd;
 @property (nonatomic, copy) NSString *adPlacementId;
 @property (nonatomic, copy) NSString *appId;
+@property (nonatomic, assign) BOOL adHasValid;
 @end
 
 @implementation PangleRewardedVideoCustomEvent
@@ -25,7 +25,7 @@
 }
 
 - (BOOL)hasAdAvailable {
-    return self.rewardVideoAd.isAdValid;
+    return self.adHasValid;
 }
 
 - (BOOL)enableAutomaticImpressionAndClickTracking {
@@ -34,6 +34,7 @@
 
 - (void)requestAdWithAdapterInfo:(NSDictionary *)info adMarkup:(NSString *)adMarkup {
     BOOL hasAdMarkup = adMarkup.length > 0;
+    self.adHasValid = NO;
     
     if (info.count == 0) {
         NSError *error = [NSError errorWithDomain:NSStringFromClass([self class])
@@ -47,13 +48,15 @@
     }
     
     self.appId = [info objectForKey:kPangleAppIdKey];
-    if (BUCheckValidString(self.appId)) {
-        [PangleAdapterConfiguration pangleSDKInitWithAppId:self.appId];
+    NSString *appIdString = self.appId;
+    if (appIdString && [appIdString isKindOfClass:[NSString class]] && appIdString.length > 0) {
+        [PangleAdapterConfiguration pangleSDKInitWithAppId:appIdString];
         [PangleAdapterConfiguration updateInitializationParameters:info];
     }
     
     self.adPlacementId = [info objectForKey:kPanglePlacementIdKey];
-    if (!BUCheckValidString(self.adPlacementId)) {
+    if (!(self.adPlacementId &&
+          [self.adPlacementId isKindOfClass:[NSString class]] && self.adPlacementId.length > 0)) {
         NSError *error = [NSError errorWithDomain:NSStringFromClass([self class])
                                              code:BUErrorCodeAdSlotEmpty
                                          userInfo:@{NSLocalizedDescriptionKey: @"Incorrect or missing Pangle placement ID. Failing ad request. Ensure the ad placement ID is correct on the MoPub dashboard."}];
@@ -66,17 +69,23 @@
     BURewardedVideoModel *model = [[BURewardedVideoModel alloc] init];
     model.userId = self.adPlacementId;
     
-    if (BUCheckValidString([PangleAdapterConfiguration userId])) {
-        model.userId = [PangleAdapterConfiguration userId];
+    NSString *userId = [PangleAdapterConfiguration userId];
+    if (userId && [userId isKindOfClass:[NSString class]] && userId.length > 0) {
+        model.userId = userId;
     }
-    if (BUCheckValidString([PangleAdapterConfiguration rewardName])) {
-        model.rewardName = [PangleAdapterConfiguration rewardName];
+    
+    NSString *rewardName = [PangleAdapterConfiguration rewardName];
+    if (rewardName && [rewardName isKindOfClass:[NSString class]] && rewardName.length > 0) {
+        model.rewardName = rewardName;
     }
+    
     if ([PangleAdapterConfiguration rewardAmount] != 0) {
         model.rewardAmount = [PangleAdapterConfiguration rewardAmount];
     }
-    if (BUCheckValidString([PangleAdapterConfiguration mediaExtra])) {
-        model.extra = [PangleAdapterConfiguration mediaExtra];
+    
+    NSString *extra = [PangleAdapterConfiguration mediaExtra];
+    if (extra && [extra isKindOfClass:[NSString class]] && extra.length > 0) {
+        model.extra = extra;
     }
     
     BURewardedVideoAd *RewardedVideoAd = [[BURewardedVideoAd alloc] initWithSlotID:self.adPlacementId rewardedVideoModel:model];
@@ -103,7 +112,7 @@
         MPLogAdEvent([MPLogEvent adShowAttemptForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
     } else {
         NSError *error = [NSError
-                          errorWithDomain:MoPubRewardedVideoAdsSDKDomain
+                          errorWithDomain:MoPubRewardedAdsSDKDomain
                           code:BUErrorCodeNERenderResultError
                           userInfo:@{NSLocalizedDescriptionKey : @"Failed to show Pangle rewarded video."}];
         MPLogAdEvent([MPLogEvent adShowFailedForAdapter:NSStringFromClass(self.class) error:error], [self getAdNetworkId]);
@@ -116,7 +125,7 @@
 
 - (void)rewardedVideoAdDidLoad:(BURewardedVideoAd *)rewardedVideoAd {
     MPLogAdEvent([MPLogEvent adLoadSuccessForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
-    
+    self.adHasValid = YES;
     [self.delegate fullscreenAdAdapterDidLoadAd:self];
 }
 
@@ -128,27 +137,23 @@
 
 - (void)rewardedVideoAdDidVisible:(BURewardedVideoAd *)rewardedVideoAd {
     MPLogAdEvent([MPLogEvent adWillAppearForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
-    [self.delegate fullscreenAdAdapterAdWillAppear:self];
+    [self.delegate fullscreenAdAdapterAdWillPresent:self];
     
     MPLogAdEvent([MPLogEvent adShowSuccessForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
     [self.delegate fullscreenAdAdapterDidTrackImpression:self];
     
     MPLogAdEvent([MPLogEvent adDidAppearForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
-    [self.delegate fullscreenAdAdapterAdDidAppear:self];
+    [self.delegate fullscreenAdAdapterAdDidPresent:self];
 }
 
 - (void)rewardedVideoAdDidClose:(BURewardedVideoAd *)rewardedVideoAd {
     MPLogAdEvent([MPLogEvent adWillDisappearForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
+    [self.delegate fullscreenAdAdapterAdWillDismiss:self];
     [self.delegate fullscreenAdAdapterAdWillDisappear:self];
     
     MPLogAdEvent([MPLogEvent adDidDisappearForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
     [self.delegate fullscreenAdAdapterAdDidDisappear:self];
-    
-    // Signal that the fullscreen ad is closing and the state should be reset.
-    // `fullscreenAdAdapterAdDidDismiss:` was introduced in MoPub SDK 5.15.0.
-    if ([self.delegate respondsToSelector:@selector(fullscreenAdAdapterAdDidDismiss:)]) {
-        [self.delegate fullscreenAdAdapterAdDidDismiss:self];
-    }
+    [self.delegate fullscreenAdAdapterAdDidDismiss:self];
 }
 
 - (void)rewardedVideoAdDidClick:(BURewardedVideoAd *)rewardedVideoAd {
@@ -171,8 +176,9 @@
 
 - (void)rewardedVideoAdServerRewardDidSucceed:(BURewardedVideoAd *)rewardedVideoAd verify:(BOOL)verify {
     if (verify) {
-        NSString *currencyType = BUCheckValidString(rewardedVideoAd.rewardedVideoModel.rewardName) ? rewardedVideoAd.rewardedVideoModel.rewardName :kMPRewardCurrencyTypeUnspecified;
-        MPRewardedVideoReward *reward = [[MPRewardedVideoReward alloc] initWithCurrencyType:currencyType amount: @(rewardedVideoAd.rewardedVideoModel.rewardAmount)];
+        NSString *rewardName = rewardedVideoAd.rewardedVideoModel.rewardName;
+        NSString *currencyType = (rewardName && [rewardName isKindOfClass:[NSString class]] && rewardName.length > 0) ? rewardName :kMPRewardCurrencyTypeUnspecified;
+        MPReward *reward = [[MPReward alloc] initWithCurrencyType:currencyType amount: @(rewardedVideoAd.rewardedVideoModel.rewardAmount)];
         
         MPLogEvent([MPLogEvent adShouldRewardUserWithReward:reward]);
         
@@ -182,12 +188,13 @@
     }
 }
 
-- (void)rewardedVideoAdServerRewardDidFail:(BURewardedVideoAd *)rewardedVideoAd {
-    MPLogInfo(@"Rewarded video ad server failed to reward.");
+- (void)rewardedVideoAdServerRewardDidFail:(BURewardedVideoAd *)rewardedVideoAd error:(nonnull NSError *)error {
+    MPLogInfo(@"Rewarded video ad server failed to reward: %@", error);
 }
 
 - (NSString *) getAdNetworkId {
-    return (BUCheckValidString(self.adPlacementId)) ? self.adPlacementId : @"";
+    NSString *adPlacementId = self.adPlacementId;
+    return (adPlacementId && [adPlacementId isKindOfClass:[NSString class]]) ? adPlacementId : @"";
 }
 
 @end