+++ Chartboost/CHANGELOG.md
@@ -1,4 +1,14 @@
 ## Changelog
+  * 8.4.1.1
+      * Add `fullscreenAdAdapterAdWillPresent` and `fullscreenAdAdapterAdDidPresent` to notify publishers of the fullscreen ad show event. Remove `fullscreenAdAdapterAdWillAppear` and  `fullscreenAdAdapterAdDidAppear` as they are now deprecated by the MoPub iOS SDK.
+      * Publishers must use v5.17.0 of the MoPub SDK at the minimum.
+
+  * 8.4.1.0
+      * This version of the adapters has been certified with Chartboost 8.4.1 and MoPub SDK 5.16.1.
+
+  * 8.4.0.1
+      * Add support for `fullscreenAdAdapterAdWillDismiss` when a fullscreen creative is about to close. Publishers must use v5.16.1 of the MoPub SDK at the minimum.
+
   * 8.4.0.0
       * This version of the adapters has been certified with Chartboost 8.4.0 and MoPub SDK 5.15.0.